requirements.txt
pygame-ce>=2.5.0
pygame_gui>=0.6.12
